chore(build): zero warnings — patch vendored jsmn, drop all -Wno-error/-Wno-* softening (#72)
* chore(build): zero warnings — patch vendored jsmn for -Wsign-conversion, drop every -Wno-error/-Wno-* softening (Power-of-10 rule 10)


* chore(build): strip the last -Wno-unused-* suppressions from core_selftest/unity_vendored; fix at source


* fix(core): gcc-only -Wsign-conversion sites in the host harness (CI host-tests red after softening removal)


---------

+++ components/core/ui/render.c
@@ -184,7 +184,10 @@ static void fb_blit_1bpp(fb_t *fb, int x, int y, int cw, int ch, int stride, con
             uint8_t ink = 0;
             if (bitmap != NULL) {
                 uint8_t byte = bitmap[(size_t)row * (size_t)stride + (size_t)(col / 8)];
-                ink = (uint8_t)((byte >> (7 - (col % 8))) & 1u);
+                /* col >= 0 by construction: cx0 >= x (clip clamps up, not down), so
+                 * col = ix - x >= cx0 - x >= 0 for every ix in [cx0, cx1). Shift in the
+                 * unsigned domain so the result stays unsigned end to end (gcc -Wsign-conversion). */
+                ink = (uint8_t)(((unsigned int)byte >> (7 - (col % 8))) & 1u); /* p10 */
             }
             fb_set_px(fb, ix, iy, ink);
         }
